+++ app/src/main/java/com/example/ymd/Home/HomeAdapter.kt
@@ -1,0 +1,4 @@
+package com.example.ymd.Home
+
+class HomeAdapter {
+}
