+++ app/src/main/java/com/example/ymd/Hot/HotAdapter.kt
@@ -1,0 +1,7 @@
+package com.example.ymd.Hot
+
+import android.content.Context
+import androidx.recyclerview.widget.RecyclerView
+
+class HotAdapter(private val mContext: Context)  {
+}
