+++ app/src/main/java/com/example/ymd/MainActivity.kt
@@ -3,11 +3,11 @@ package com.example.ymd
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.viewpager2.widget.ViewPager2
+import com.example.ymd.Home.HomeFragment
+import com.example.ymd.Hot.HotFragment
+import com.example.ymd.Mypage.MypageFragment
 import com.example.ymd.databinding.ActivityMainBinding
-import com.example.ymd.databinding.FragmentHomeBinding
-import com.example.ymd.databinding.FragmentHotBinding
-import com.example.ymd.databinding.FragmentMypageBinding
-import com.example.ymd.videoList.SearchFragment
+import com.example.ymd.Search.SearchFragment
 import com.google.android.material.tabs.TabLayoutMediator
 
 class MainActivity : AppCompatActivity() {
@@ -46,9 +46,9 @@ class MainActivity : AppCompatActivity() {
 
         TabLayoutMediator(binding.tableLayout, binding.mainViewpager) { tab, position ->
             when (position) {
-                0 -> tab.setIcon(R.drawable.home).text = "Home"
-                1 -> tab.setIcon(R.drawable.hot).text = "Hot"
-                2 -> tab.setIcon(R.drawable.mymedia).text = "Mypage"
+                0 -> tab.setIcon(R.drawable.home).text = "com/example/ymd/Home"
+                1 -> tab.setIcon(R.drawable.hot).text = "com/example/ymd/Hot"
+                2 -> tab.setIcon(R.drawable.mymedia).text = "com/example/ymd/Mypage"
                 3 -> tab.setIcon(R.drawable.search).text = "Search"
             }
         }.attach()
