+++ app/src/main/java/com/example/ymd/Mypage/MypageAdapter.kt
@@ -1,0 +1,4 @@
+package com.example.ymd.Mypage
+
+class MypageAdapter {
+}
