+++ app/src/main/java/com/example/ymd/Mypage/MypageFragment.kt
@@ -1,0 +1,20 @@
+package com.example.ymd.Mypage
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.ymd.R
+
+class MypageFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_mypage, container, false)
+    }
+
+}
