+++ app/src/main/java/com/example/ymd/datail/DetailAdapter.kt
@@ -1,0 +1,4 @@
+package com.example.ymd.datail
+
+class DetailAdapter {
+}
