+++ app/src/main/res/layout/activity_detail.xml
@@ -4,7 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".VideoDetailFragment"
     android:padding="20dp">
