+++ app/src/main/res/layout/fragment_search.xml
@@ -5,7 +5,7 @@
     android:id="@+id/searchFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".videoList.SearchFragment">
+    tools:context=".Search.SearchFragment">
 
         <ImageView
             android:id="@+id/imageView4"
